Remove Width & Radius From Flash Messages
Remove both the width and the radius of the flash messages because the
width should now go to the edge of the screen. Since the width goes to
the edge of the screen a round corner no longer is looks good since
the edge should look like its coming off the screen.