Add Provident User From Entering Blanks
Add validation to provident the User from entering blanks into the Reset
Password form because they should not be doing so. Also this will make
the site hard to use since they having if they do not enter a password
then they would have to do the reset password again.